poa: fused int16 channelizer (drop fbuf + deinterleave pass) + accumulators
poa_frontend_feed now takes int16 IQ directly (the native resampled stream)
instead of float, and consume_block reads whole rtlMult-blocks straight from
that buffer, converting to float once per sample on the stack. This removes
two full-rate passes over the 2.5 MSPS stream and a 32 KB static fbuf: the
caller's int16->float conversion and this file's old float deinterleave copy.
The mix also now uses four accumulator chains (2-way unroll x re/im) to hide
FPU add latency on the serial integrate-dump.

Correctness gate holds: host test_poa_frontend still reproduces the acarsdec
oracle (JQ0404) byte-exact and the full host suite is 75/75. The int16
interface is more faithful to the source data than the prior float path.

Measured on-device effect is modest: consumed rate 0.64 -> 0.70 MB/s, dsp still
~100%. This is a correctness/structure change; it does NOT resolve the POA DSP
saturation. Root-causing the remaining gap (channelizer still consumes only
~15% of the 2.5 MSPS stream) is the next step and is not claimed here.

Iridium/VDL2 unaffected: only the POA (p->poa_fe) branch of dsp_processor_feed
changed; the tagger memcpy path is untouched.

Smoke-skip: POA-only path (Iridium/VDL2 branches unchanged); host JQ0404
bit-exact + 75/75 + live device boot verified. VHF antenna required for POA
device smoke.

+++ common/poa_decoder/poa_frontend.h
@@ -6,8 +6,8 @@
 // output is an AM envelope, the per-block oscillator reset is harmless (plan).
 //
 // Pure C11 + libm — links into tests/host directly. Input scale is irrelevant
-// (the demod normalizes per bit), so feed float IQ of any scale (cu8-127.37
-// on device, or s16 from a capture in tests).
+// (the demod normalizes per bit), so feed int16 IQ of any scale (the device's
+// resampled s16 stream, or s16 read straight from a capture in tests).
 
 #pragma once
 
@@ -24,7 +24,9 @@ poa_frontend_t *poa_frontend_create(uint32_t fs_hz, uint32_t lo_hz,
                                     const uint32_t *chan_hz, int nch,
                                     poa_block_cb cb, void *user);
 
-// Feed nsamp COMPLEX samples as interleaved float I/Q (length 2*nsamp).
-void poa_frontend_feed(poa_frontend_t *fe, const float *iq, int nsamp);
+// Feed nsamp COMPLEX samples as interleaved int16 I/Q (length 2*nsamp).
+// The channelizer reads whole rtlMult-blocks straight from this buffer (no
+// intermediate float copy) and converts to float inline in the hot loop.
+void poa_frontend_feed(poa_frontend_t *fe, const int16_t *iq, int nsamp);
 
 void poa_frontend_destroy(poa_frontend_t *fe);
